requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.12.0=py37h06a4308_0
aiohttp=3.6.3=py37h7b6447c_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py37h06a4308_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_1
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
chardet=3.0.4=py37h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.6=py37hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py37_0
cython=0.29.22=py37h2531618_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
expat=2.2.10=he6710b0_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=pyh9f0ad1d_0
glib=2.67.4=h36276a3_1
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
google-auth=1.28.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.3=pyhd3eb1b0_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=nompi_py37hf7afa78_105
hdf5=1.10.6=nompi_h3c11f04_101
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py37h06a4308_1
intel-openmp=2020.2=254
ipykernel=5.5.0=pypi_0
ipython=7.21.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py37h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libiconv=1.15=h63c8f33_5
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py37h06a4308_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=4.7.6=py37h7b6447c_1
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.3.0=py_0
pandas=1.2.3=py37ha9443f7_0
parso=0.8.1=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.1.2=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.14.0=py37h2531618_1
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.8.1=pypi_0
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd8ed1ab_0
python_abi=3.7=1_cp37m
pytorch=1.8.0=py3.7_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.0.3=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.1=py37ha9443f7_0
scipy=1.6.1=py37h91f5cce_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.35.2=hdfb4753_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=mkl_py37h2d14ff2_0
tensorflow-base=2.4.1=mkl_py37h43e0292_0
tensorflow-estimator=2.4.1=pyheb71bc4_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.9.0=py37_cu102
tornado=6.1=py37h27cfd23_0
traitlets=5.0.5=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py37h4abf009_2
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
